requirements.txt
django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
psycopg2-binary==2.9.3
django_filter==21.1